Refactor: SetPicture Layer Abstraction
