+++ backend/src/models/DataApiResponse.ts
@@ -2,3 +2,8 @@ export interface DataApiResponse {
   dbResponse: AWS.RDSDataService.ExecuteStatementResponse;
   error: string;
 }
+
+export interface SetPictureResponse {
+  result: DataApiResponse;
+  imageUrl: string;
+}
